Improve support for intercepting RenderContext calls.
Instead of handling a `BaseRenderContext` directly, `WorkflowInterceptor.onRender` never
sees the `BaseRenderContext` itself but can pass a new type, `RenderContextInterceptor`,
to its `proceed` function. `RenderContextInterceptor` is an interface similar to `WorkflowInterceptor`,
with a bunch of methods that mirror `BaseRenderContext` but each have their own `proceed` function.
This makes it much simpler to intercept calls to `actionSink.send` and to intercept side effects
when they're actually ran as opposed to when they're requested.

Also uses the new class in `SimpleLoggingWorkflowInterceptor` to add additional logging.

Fixes #330 and fixes #331.